Experiment: generate an image of the current schedule
The image is generated as an SVG, encoded as a URL, and exposed as an attribute to each sensor.

The image can be displayed using, eg:

type: custom:button-card
entity: sensor.festive_lights_state
show_entity_picture: true
entity_picture: |
  [[[ return states['sensor.festive_lights_state'].attributes.schedule_image
  ]]]
size: 90%
styles:
  card: null
  img_cell:
    - padding-left: 10px
    - padding-right: 10px

Can't get the SVG titles to show up. Maybe because this is embedded as an image?