Issue 92 (#96)
* fixes #92

Templates in extra attributes were being evaluated every minute (on each update),
instead of at every `refresh` interval. This resulted in the template being used
as the key in an internal data structure instead of the template result. The problem
is that HA Templates can't be used as keys, so the internal data structure was
getting corrupted.

* bump version to 0.18.1